stream: Move requests session helper to internal
