+++ mopidy/internal/http.py
@@ -1,0 +1,16 @@
+from __future__ import absolute_import, unicode_literals
+
+import requests
+
+from mopidy import httpclient
+
+
+def get_requests_session(proxy_config, user_agent):
+    proxy = httpclient.format_proxy(proxy_config)
+    full_user_agent = httpclient.format_user_agent(user_agent)
+
+    session = requests.Session()
+    session.proxies.update({'http': proxy, 'https': proxy})
+    session.headers.update({'user-agent': full_user_agent})
+
+    return session
